src/useCases/accounts/validateCheckIn/validateCheckIn.spec.ts
import { InMemoryCheckInsRepository } from '@/repositories/prisma/in-memory/in-memory-checkIn-repository'
import { ValidateCheckInUseCase } from './validateCheckIn'
import { beforeEach, describe, it, expect } from 'vitest'
import { ResourceNotFound } from '../errors/resourceNotFound'

let checkInsRepository: InMemoryCheckInsRepository
let sut: ValidateCheckInUseCase

describe('Check-In Use Case', () => {
  beforeEach(async () => {
    checkInsRepository = new InMemoryCheckInsRepository()
    sut = new ValidateCheckInUseCase(checkInsRepository)
    // vi.useFakeTimers()
  })

  // depois de executar os testes volte para as datas reais
  // afterEach(() => {
  //   vi.useRealTimers()
  // })

  it('should be able to valitade the check-in', async () => {
    const createdCheckIn = await checkInsRepository.create({
      gym_id: 'gym-01',
      user_id: 'user-01',
    })

    const { checkIn } = await sut.execute({
      checkInId: createdCheckIn.id,
    })

    // espero que o id do usuario seja do tipo de qualquer string.
    expect(checkIn.validated_at).toEqual(expect.any(Date))
    expect(checkInsRepository.items[0].validated_at).toEqual(expect.any(Date))
  })

  it('should not be able to valitade an inexistent check-in', async () => {
    await expect(() =>
      sut.execute({
        checkInId: 'inexistent-check-in-id',
      }),
    ).rejects.toBeInstanceOf(ResourceNotFound)
  })
})
